+++ etherlink/kernel_latest/runtime/src/safe_storage.rs
@@ -201,6 +201,15 @@ impl<Host: StorageV1> StorageV1 for SafeStorage<&mut Host> {
         let path = safe_path(path)?;
         self.host.store_value_size(&path)
     }
+
+    #[inline(always)]
+    fn store_get_hash(
+        &self,
+        path: &impl Path,
+    ) -> Result<[u8; tezos_smart_rollup_core::STORE_HASH_SIZE], RuntimeError> {
+        let path = safe_path(path)?;
+        self.host.store_get_hash(&path)
+    }
 }
 
 impl<Host: WasmHost> WasmHost for SafeStorage<&mut Host> {