+++ src/kernel_sdk/core/src/rollup_host.rs
@@ -138,4 +138,14 @@ unsafe impl SmartRollupCore for RollupHost {
     unsafe fn reveal_metadata(&self, destination_addr: *mut u8, max_bytes: usize) -> i32 {
         target_impl::reveal_metadata(destination_addr, max_bytes)
     }
+
+    unsafe fn __internal_store_get_hash(
+        &self,
+        path: *const u8,
+        path_len: usize,
+        destination_addr: *mut u8,
+        max_size: usize,
+    ) -> i32 {
+        target_impl::__internal_store_get_hash(path, path_len, destination_addr, max_size)
+    }
 }